refactor: split major views into components
